Break test helpers into their own packages
